Create wrapper.py
This script assists in attaching a Frida script to applications that encounter the "Failed to spawn: unable to find front-door activity" error. Script is continuously checking for the application to be manually opened by the user (via app drawer or adb). Once the app is detected as running, Frida script is attached.